backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
databases
websockets
python-dotenv
psycopg2-binary